Instead of prohibiting ROW in a stanza of REFERENCING, require it.
The standard syntax would require making the keyword ROW at least
partially reserved.  To avoid that ROW must be prohibited in the
REFERENCING clause, required for ROW-oriented definition in the
REFERENCING clause, or the AS keyword in the REFERENCING clause
must not be optional.  Requiring ROW for a row definition seems
cleanest and easiest to explain.

This is all academic, though, since at this time we are throwing
an ERROR for row specifications; but we don't want to forclose the
possibility of later supporting standard syntax for that without
breaking working code.