Eliminate TuplestoreRelation node type from analyzed parse tree.
The idea was to capture the fact that this was not a relation from
the catalogs in the plan analysis, but there was nothing the node
type did which wasn't already covered by the RangeVar and
RangeTableEntry nodes.  If someone wants to know whether the
relation is a tuplestore, they can look at those.